Change shorthand strategy for frame fields in evaluator
In order to be able to reliably swap the frame "f" from one value to
another inside of a non-recursive evaluator, the SHORTHAND() method
which was assuming a constant `f` for an entire call to the evaluator
core routine can't be used.  This switches to a more conventional
macro so that `f->feed->xxx` becomes `f_yyy` where yyy is the renamed
field from the point of view of the evaluator.