Improve comment formatting in evaluator
Ren-C has adopted a "2 space convention for same-line comments"

     foo(); // not this

     foo();  // like this

The convention has come about as being preferred slowly, and has not
had a system-wide rewrite to enforce.  But %c-eval.c is a particularly
"flagship" file, so it should be showing best practices.

This goes ahead and makes sure it does that, rather than weaving in
the changes with other commits so they look random.