Add stackless handling to GROUP! and function arguments
This is a rough attempt at making it so that function arguments and
GROUP! evaluations in the central evaluator do not deepen the C stack.
So no more C stack is used for:

    ((((((((1 + 2))))))))

Than to run plain `null`.  The stack used is only the stack for one
function call (function calls themselves are not stackless, only the
step of aquiring arguments).

This involved needing to change tests: there were two tests of
GROUP! that were nearly identical, that started infinite looping,
where before they would have generated stack overflows.  This unifies
them into one test that just uses a counter and shows the counter
permits very deep recursion.

Includes minor improvement to ARRAY made while working on the tests.