Switch WHILE to a "stackless" implementation
This makes use of the ability to catch throws in continuations to turn
WHILE into something of a state machine, that is able to run the
condition and the branch while it is not itself on the stack.

Accomplishing this uses the "spare" cell as a state variable to say
whether the condition or the body was what was run last.  It is END
on the first run so the WHILE knows how to initialize the result for
if the body is never run (e.g. BLANK!), and then FALSE if the condition
was the last continuation, TRUE if the body was the last continuation.

The RETURN slot is only used for typechecking in natives, so the arg
position in the frame is available for other use.  This uses that for
caching the body result to be returned.  Unfortunately that loses an
optimization previously available since each evaluation could point
to a different output slot it targeted... but continuations assume
you always want to write to the same output slot that the frame had
to begin with.  This may be something to review as continuations mature.