Change NEXT_ARG_FROM_OUT from eval flag to feed flag
Due to runing out of EVAL_FLAG_XXX, the flag for getting the next
argument from the output is moved from the frame onto the feed.

Conceptually it needs to be monitored that it doesn't get mistaken
as being used for two different frames ideas of f->out.  But that
monitoring was already in place.