Make ACTION! continuations actually stackless
The first implementation of continuations was taken from the code in
Do_Branch_With_Throws(), which called into the evaluator in the normal
stack-oriented way.  First BLOCK! was stackless, then FRAME!, etc.

This takes another step by making ACTION! continuations stackless also.
It's useful for control constructs like IF that have branches which
are actions, but also enables routines like ENCLOSE to run their
functions (which are parameterized by frames) in a stackless way that
can work with generators.