Transition frames to be marked using queueing
With stackless processing, REBFRM* are dynamically allocated and can
be disconnected and reconnected into the stack in scenarios like a
generator with `yield`.  If that happens, then the GC has to mark a
REBFRM* that's not currently hooked into the execution stack.

Doing that requires the ability to mark frames that aren't hooked
into the stack reachable through FS_TOP.  This teases out the marking
so that it is recursive in following the f->prior link, and stops
when it hits the bottom (if linked into the stack) or a nullptr
(if it's a fragment).  Then goes through the usual queueing procedure
so that each recursion must queue--which prevents a cell from leading
to a possibly arbitrarily deep C stack to mark it.

It also takes a case mentioned in the frame marking code and makes it
a test vs. a comment.