Transition feeds and frames to using pooled memory
With frames being detachable from the stack for later resumption, it
becomes necessary to enumerate the existing frames for clearing their
marked-as-used bits and GC'ing frames that have expired.  The best
way of doing that is to use pooled memory--which also helps performance
with fast fixed-size allocations.

So long as this was being done for frames, it is done for feeds as
well...which obsoletes a feed leak-detection mechanism that existed
previously (since the pool can be enumerated).