Add missing EVAL_MASK_DEFAULT cases
EVAL_MASK_DEFAULT is used in case there are flags which are better
defaulted to as being 1 instead of 0 by default.  There was a debug
flag that was added to all EVAL_MASK_DEFAULT usages and tested for, but
wasn't getting checked in path-oriented frames.

This adds the missing cases, and also shifts from using the debug
flag to just testing for the flags expected to be set.

It also discontinues the "endlike" nature of the frame spare cell,
in anticipation of moving the frame flags to the head of the REBFRM.
That removes the need for the safeguard of the "not a cell" nature
of the flags (and they will need to be a "cell" in the near term for
the previously-interpreted-as-cell head of a frame to look like it).