Move flags to lead position in Reb_Frame
With Reb_Frame becoming a genuine "node" in the pooled sense, the idea
of using the frame flags as the implicit terminator for the frame's
spare cell seems like a pointless optimization.  This paves the way
for REBFRM being recognized as a distinct node type, and keeps the
spare cell aligned on a 64-bit boundary on 32-bit systems by moving
the prior frame pointer into the second platform-pointer sized slot.

Since this obsoletes the idea of the flags serving as an endlike
header for the spare cell in the frame, that feature is eliminated...
which allows the flags in the low byte to be used for typical node
purposes, and the flags in the end slot to be retaken.