Eliminate cached per-C-stack-level TG_Tick
When there were many recursive calls to evaluator functions, there
was a convenience factor to being able to look at the stack level you
were on and see a tick number reflecting the capture of what was in
the frame at that level.  But with no recursions, the cached tick
will always match TG_Tick...you would have to look in the frame stack
to get tick counts.

Considering it's not hard to look at `f->tick` or `TG_Tick` if you
want to, this convenience of diminishing value is removed.