Factor Eval_Arg() out to a separate function
As a step toward making argument processing its own state machine
step (and losing the idea of a controlling `while()` loop that has
to have a `goto` into it to resume on argument fulfillment), this
separates Eval_Arg() into a function.

It is still inside a while loop, but the return states are made more
clear, with internal gotos separated from the "external" ones.