Homogenize R_CONTINUATION to a common "frame language"
As the workings of stackless continue to sift out, the general
idea seems to be that all continuations are based on building a
frame as the currency for the continuation and linking it in...
then getting an optional callback.

Initial concepts were that this would be memoized in some way that the
dispatcher could help you with.  This change transitions to where
there is no memoization outside of the state of the frame itself;
which means R_CONTINUATION always means "return to the root frame
evaluation loop".