Eliminate enfix-based REEVALUATE case
The driver for running continuing frames was made more complicated by
handling a form of enfix dispatch, which could be simulated by doing
a normal continuation frame.  That code was necessary for boot in
SHOVE, so the best short-term change seemed to be to go ahead and
make SHOVE use a continuation and do the enfix case on its own.

This simplifies the top-level eval loop, in a way that is intended
to prep it for a more extensible (e.g. not EVAL_FLAG-based) way to
get to the evaluator routine that is desired.