Workaround for `initial_flags` capture settings
Frames were made to be reused cheaply, e.g. for successive evaluations.
To get this feature, it was considered desirable that they return to
their initial flag state by the end.  Even though it's easy to reset
all the flags at once, the idea would be that stray flags would
be a sign of some unintended effect...so they should be zeroed out
(at least in debug builds).

This idea is somewhat antiquated--and resetting the flags on each
frame reuse with selective debug checks for leaked flags is probably
the better idea.  However, the idea had caught some bugs; so a more
formal framing of an alternative is probably a good idea before just
throwing it out.

This does a small patch to the concept to get the tests running in the
new stackless model.