Very creaky (but booting) hack for functions own frames
This breaks function calls out to having their own frame identity, as
far as the evaluator loop is concerned.  This potentially paves the
way for a FRAME! object representing things like a GROUP! or BLOCK!
evaluation.  The thing to like about the idea of a concrete Rebol
type to represent such evaluations is that it would be usable as
state, e.g. for natives like REDUCE...which need to step through
efficiently while disallowing modifications.

This was largely guesswork, trying to achieve the physical outcome of
a separate frame for function invocations.  Performing this separation
broke several assumptions, such as that EVAL_FLAG_FULFILLING_ARG would
imply in non-variadic situations that the frame above was a function
frame that was also executing.

Hence a broad review of the implications of this experiment is needed,
if the idea is that a central evaluator loop is going to control the
frame evaluation and continuation process.