Move "trampoline" into its own file
In researching what it would take to implement stackless, it's clear
that there needs to be a single routine which loops over whatever the
top of the stack becomes.  Converting the entire system to work in
terms of this approach is a large task; in particular figuring out how
things that were formerly framed in terms of the stack (throws, longjmp)
fit into such a new paradigm.

For the sake of not inventing new terminology for something that already
exists elsewhere, this adopts the Lisp term for calling this master
loop a "trampoline".  It breaks it out into its own file to help it
be seen and understood more clearly.