Basic first cut at a "coherent" Trampoline
Boots and shuts down successfully (but does not run tests).

This moves to a model where the trampoline will just keep calling a
frame on the top of the stack until its "executor" function is nullptr.
It's a shift away from the idea of controlling such behavior with
flags (which could get out of sync with an executor), and hence
eliminates EVAL_FLAG_CONTINUATION -- which was introduced at a very
early moment in the process.

Similarly, it switches over to using a frame `arg` value of nullptr
to indicate that a frame hasn't started enumerating yet...and a
`param` value of END to indicate that enumeration is done and it
is thus in some kind of post-processing phase for the function.