Use 8 bits of eval flags for a "state byte"
The frame's "spare" cell was being used as state for deciding where
to `goto` in natives.  The logistics of these states didn't make much
sense, and sacrificed a whole GC-capable cell (which could be used
as an evaluation target, with some modification).

This switches over to a much more conventional way of using a byte
out of the eval flags in order to decide where to jump.  Then IF,
CASE2, WHILE, and REDUCE2 are converted to the more readable forms.