Parameterize trampoline with "root frame"
Initially the assumption was that the trampoline would stop running
if it ever came back to find the frame that was the top of the stack
when it started being in the "completed" state.  This allows the
trampoline to be parameterized with a different concept of what the
completion goal is.

The reason to do this is that it permits a stack of frames to be lined
up in advance and then set into motion.  This could permit something
like an API for continuations that could be used by libRebol--capturing
a stack and then restoring it from C calls.