Use continuation branch code path for Do_Branch_Throws()
The first thing that was converted to stacklessness was IF, which
meant being able to run the various branch forms as continuations.
The API picked for that was to be able to `return Init_Continuation(`
which packaged together the idea of pushing a new frame to execute
along with returning R_CONTINUATION.

One idea that was implicit in this was that the return code could opt
to be simply a value, instead of a continuation (e.g. a quoted branch
could just dequote the branch and not be a frame).  There could also be
an R_THROWN.

This didn't account for the idea that a function dispatcher might be
set up in a generic way (as IF was) to want to be called back, even
if the branch was finished.  That might mean doing some post-processing
(such as voidification of nulls), or moving on to a next phase (as
in the CASE construct).

In an effort to eliminate the separate code path for Do_Branch_XXX(),
this separates out the continuation so that it is always assumed that
a frame gets pushed.  It could be optimized in the future, but for
the sake of simplicity it will push a frame...even if the frame just
says "use what's in the output".