Get rid of DEBUG_ENSURE_FRAME_EVALUATES
When the question of making debug steps interceptible came up prior
to stackless, a flag called DEBUG_ENSURE_FRAME_EVALUATES was added.
It didn't work out particularly well, but was kept around as a reminder
of the importance of making sure whatever evaluation mechanic actually
got called and didn't subvert it (as had created loops in R3-Alpha
you couldn't Ctrl-C out of, like `forever []`).

Stackless is going down a different path of assurance, by way of the
"trampoline" function that gains control on every evaluation.  The
Ctrl-C handling--as well as debugging hook points--are done through
state transitions there.

So this cleans out the flag (which was disabled anyway).