Make subrule f->param, put "save" into frame
This takes two more pieces of state that were local and turns them
into persistent forms.  The "save" cell is used as the backing store
for word fetches that may be unstable (e.g. if a rule is in a variable
that might be changed during the execution of a GROUP! in the rule).
And the "subrule" pointer, which might either be into the save cell
or into somewhere in the actual rules block.

The subrule pointer uses the f->param pointer in the frame, to help
avoid "unnecessary derelativization".  This is currently legal as there
is no associated ACTION! running.