P_RULE => P_RAW_RULE, P_RULE for `rule` local
In continuing to try and stop the use of C local variables for state
inside of PARSE, this switches the `rule` local variable to be P_RULE
that actually uses a frame pointer variable to hold its state.

Both P_RULE and P_SUBRULE are "optimized" in the sense that they do not
needlessly derelativize and move cells that are held locked in the
rule frame.  They will point directly to those locations, but will
point to a temporary location where variables have been fetched or
paths have been evaluated.

(It's not clear how the rule/P_RULE variable has avoided contention
with the "subrule" over use of `spare`, but that should be looked at.)