Make parse flags persistent, unify find and state flags
R3-Alpha's parse had several notions of "flags"; notably one called
`enum parse_flags` and the other called `enum Parse_Flags`.  The
former would be accessed in a parse state structure as `parse->flags`,
while the latter would be available as a local in the main parse
function as `flags`.

Adding to the confusion, the `enum parse_flags` had a mysterious
overlap with the AM_FIND_XXX flags.  Due to these confusions, Ren-C
added a flag for determining if one rule was to be executed vs.
many into the Parse_Flags, and OR'd it with the parse_flags.

This sifts out the issue a bit by making the relationship to the
AM_FIND_XXX flags explicit through a static assert.  It takes the
question of whether one rule or many should be executed and unifies
that with the EVAL_FLAG_TO_END--already used for the same meaning
in the plain evaluator (e.g. "DO/NEXT" vs. plain DO).