Make BLOCK!-subrules in PARSE stackless
This makes plain recursions via BLOCK! (not INTO operations) act as
stackless.  Since previous recursions did not actually use the
dispatcher, it was not noticed that they were doing COLLECT rollback
based on still having access to frames that were thrown.  This thus
also had to move the collect rollbacks into the parse executor itself.

Another required change (given how the code is written) is that the
SET_OR_COPY_WORD be folded into the frame state.  This could be changed
to be done with another mechanism, but it's just put into a misc
frame field for now.