Initial implementation of stackless COMPOSE
This addresses the issue of a recursive native that wants to be
stackless.  It uses frames to hold the recursion state, which is
processed via the Trampoline without each recursion using the C
stack.

Rather than making COMPOSE itself recurse, a new frame "executor"
(a Composer_Executor()) is used.  This puts it in the same stack as
ordinary function calls, but with a distinct frame processor.  This
implements an idea which doesn't try to use much state in the
"composer" frames by not having "varlists" of their own.  Instead
they are able to refer back to the original COMPOSE frame via putting
a FRAME! value in the spare cell--since ACTION!-bearing FRAME!s can
be "reified" in that fashion.

This is distinct from the recursion method for PARSE, which leaves
the door open to having varlists for each SUBPARSE recursion.