Undo breaking of Arg_Executor() out from Action_Executor()
It wasn't clear in the beginning of the stackless breakdown if there
would need to be some kind of separate frame to use for arguments.
It also seemed like factoring out the idea of processing a single
argument out of a loop was likely to make things clearer regardless.

In practice, separating the arguments from the loop created a confusing
set of questions, and wound up making things worse.  Factorization in
another way may be better, as there was no design requirement to make
new frames with new executors for the arguments.

This backs out that change for now, so the code can be studied more
holistically.