Use state byte in action dispatcher, fix GC bug
There was no "state byte" initially, and so determining special modes
was done with flags.  The desire to stop consuming arbitrary frame
flags led to an attempt to determine if the arguments were being
fulfilled or not (or other mode questions) by manipulating the argument
state to be nullptr on initial entry and then set to the args head
inside the action executor.

This led to problems due to breaking some invariants needed e.g. by
the GC when it happened at certain times.  Emergence of the state byte
gave another possibility, as the byte is available for use when not
being used by an action dispatch phase.  This allowed the code for
initializing f->arg to return to its former/saner way.

This additionally folds the before/after action checks into the action
exeecutor instead of the independent %d-eval.c file.  That file was
made when the action dispatch was living in the same file as the
general evaluator, making it overlong.  Given the small size of the
existing checks, it made more sense to have them included so they
would not be out of sight or mind (e.g. redundant checks on f->arg
being FRM_ARGS_HEAD() on process_action entry).

Also, a subtle GC bug can be addressed with this change...where a
frame's arguments which had no other reference were being GC'd in
ENCLOSE of a function.