Don't overwrite cell-formatted tick during cell copy
The existing tick in a formatted cell says where that cell came from
(e.g. the tick at the time of Copy_Array(), or Init_XXX(), etc.)
While it might seem like copying that tick from the source cell in a
move would help you trace the "true origin of the value", that is
already covered most of the time by the tick in the series node both
cells will point at.  Hence more useful to leave the destination alone.

As a compromise in case there's something about the "originating cell"
that is being tracked, the "touch" tick is copied in moves.