Choose state names other than GOTO labels
The initial concept of states was that they would match the name of
the `goto` they jumped to.  This turns out to be less coherent than
allowing the state byte to describe the intent or meaning of the
current continuation.

In addition, blockscope was used alongside the labels.  The original
creation of blockscope was to help comment why the braces were there.
But when a label is at the location, the reasoning is clear.  This
deletes those blockscope references.