Patch GC problem in ENCLOSE
ENCLOSE was using a "with" parameter to a continuation that was not
being GC protected.  This stows it in the FRM_SPARE() slot.

A more optimized concept is suggested which would fill the frame
of the function by looking for the first unspecialized argument.  This
would avoid needing to allocate a feed.