Make DO of string and block stackless, COMPOSE2 => COMPOSE
This brings some amount of stacklessness to the DO operation by making
it stackless for string and block/group executions.  The variadic
code is not yet stackless.  String executions build an array to call
the `DO*` system helper, which is less efficient than directly making
a frame...but there's not a convenient way to do that from C yet.

It also promotes the experimental stackless COMPOSE2 to be the main
COMPOSE.