Add comments on why mutable_FIRST_BYTE() etc. are needed
A mistaken notion that FIRST_BYTE() could be replaced with the
definition of mutable_FIRST_BYTE() led to an ill-advised attempt to
unify them without realizing that would not work (due to const data
in `const REBVAL*` etc.

This adds a comment about it, but does make QUOTING_BYTE() not need
the mutable_QUOTING_BYTE() case.