ENCLOSE + action branches w/arg faster, more rigorous
The methodology by which a single parameter was passed to a function
was based on pushing the action frame and then making a single-item
feed to process.  This meant the quoting convention of the function
came into play when feeding it a singular argument, as well as costing
the overhead of a feed.  Plus, the value being fed in had to be in a
GC safe position for the next trampoline evaluation.

This commit sidesteps those issues by creating a frame for the action
directly, and locating its first unspecialized argument (if any), and
writing the argument directly into the frame.  This means that the
argument gets typechecked without having to go through any parameter
conventions, and no feed is needed.