EVAL_FLAG_ROOT_FRAME on non-unwindable frames
When a Trampoline instance is invoked from the C stack, at present
that cannot be unwound past.  (Being able to do so would involve the
caller being able to handle something akin to a "Promise" as a result
instead of an actual value, that would package up the stack.)  It's
important to flag these frames so that YIELD can recognize them and
raise an error if one is hit before being able to reach the generator
or yielder that they are bound to.

Because all of a frame's flags (including the STATE_BYTE) can be
initialized in one assignment to a header-sized variable, it is more
efficient to have this flag put on by the caller of the Trampoline
instead of being done by the trampoline itself.  Hence it only checks.

At the moment, there are quite a number of un-converted "stackful"
calls into the Trampoline.  This adds the flags at those callsites,
which were found by basically running all the tests to see when the
Trampoline would complain.