Make REBFRM track the API handles that are used
Initially, API handles would point to a reified FRAME! that owned the
handle.  Leaks would only be noticed when a GC ran, and the frame had
to track a state indicating failure so that handles could be forgiven
for being leaked.

There were several disadvantages of that approach.  Firstly--it meant
finding out about leaks a bit late.  Secondly, only frames that were
running functions could accrue API handles...and those frames were
forced into reification so that the handles could point at them.  This
meant creating reified varlists for every API call.  Also, "dummy
functions" had to be pushed even when they were not necessary.

This takes the hit of increasing the frame size by one pointer to
track a list of extant API handles.  The handles are in a doubly-linked
list in order to make their allocation and freeing fast, given that
many may be allocated and there is no guarantee of a "stacklike"
ordering.