Eliminate the global "Dummy_Action" and uses
Because API handles could only be attached to action frames (with
reified varlists), when a frame was needed to attach API handles to
then not only did a "dummy frame" have to be created, but a "dummy
action" had to be pushed...and dropped.

In addition to costing more than a simple frame, this complicated a
number of mechanisms that had to filter out the dummy frame as not
being a legitimate function call.

Switching to allowing frames of any kind to track the list of API
handles that are allocated while they are the top frame makes this
unnecessary, so all of it is deleted.