Eliminate VARLIST_FRAME_FAILED
The notion of flagging varlists that were killed off by a fail() is
something that isn't necessary for the purpose it was introduced any
longer.  That was to let the GC know when it was okay to free API
handles that hadn't been released.  Now that arbitrary frames track
their own handles, the freeing can be managed in a more precise way.