Permit GC during thrown states
There was a kind of arbitrary choice to not let the garbage collector
run during a throw, which had been put in place at some point.  It's
not clear that has any particularly good reason...and it limits the
places you can do a RECYCLE/TORTURE-style tests.  This adds ability
to recycle during a throw.