Add Yielder and Generator tests to %core-tests.r
During the earliest phases of stackless transition, experiments
with yielders and generators were a way of informing the design--
even though they were too complex to work reasonably.

As stackless has become a bit more mature, these constructs have
approached a level of stability that they can now be included in
the test suite...without generating asserts or leaks on shutdown.
(Not that it would be a great idea to start using these in
bootstrap yet.  But they're starting to be all right.)

This goes ahead and adds in tests for generators and yielders,
as well as moving the old usermode GENERATE out into its test
file.  Equivalent tests for the new generator for each GENERATE
tests are included.

(Includes some tests that were found in random scratch files.)