Stackless FOR (no series iter), eliminate REPEAT
This is an inefficient implementation of FOR for starters, which avoids
the bit-level optimization in favor of leaning on Rebol's existing
generic comparisons and arithmetic.

It's slow for now because it goes through the libRebol API...but this
is good enough to prototype the desired functionality and do it in a
stackless way.  Leaning on these mechanisms incentivizes making them
faster (which can be approached several ways...for starters by being
able to reference generics by their pointers instead of having to
transcode and bind them.)

FOR had been able to iterate over series as well as the conventional
numeric ranges--though this was not used often.  So the series
functionality was removed and the tests put on a forum post:

https://forum.rebol.info/t/generic-looping-cutting-for-back-a-bit/1310

Also, rather than worry about the details of implementing REPEAT in
a stackless way this simply removes it.  That's because its function
was covered by FOR-NEXT and COUNT-UP, which are more descriptive...
and REPEAT is a better candidate for doing what LOOP does today.