Simplify CONSOLE through goroutine use
In its initial conception, the CONSOLE sought to remove itself
from the stack during evaluations in order to have the only frames
be those of the code being evaluated.  This kept the console from
showing up in the stacktrace...but it also permitted a C level
"driver" loop to distinguish between when the console itself was
running (and Ctrl-C should be disabled) and when user code was
running.  It also distinguished when code in service to the
console itself (e.g. skin code) was running, where Ctrl-C would
lead to reverting to the default skin.

This simplifies that using goroutines and channels, so that the
console can actually be formed as a "read, eval, print loop" that
uses ordinary looping (CYCLE) and maintains its step in the state
based on its stack.  It gives up control by way of sending on
a channel the request it wnats the C driver to fulfill, and then
blocks on receiving it.

In moving to this simplification, the idea that the startup
sequence offloads code to the console to set up proper Ctrl-C
handling is eliminated.  The new rule is that it should be
possible to make a build that has command line processing but
no console.  This will necessitate a new kind of interrupt
handling strategy, which is being designed next.