Execute Usermode Code on separate Goroutine
This introduces the notion of /KERNEL goroutines, and begins on an idea
that usermode code would not be running on the same stack as the
console itself.  The granularity of cancellability (Ctrl-C) and of
debuggability would then be properties of the individual tasks.

(While not necessarily the best idea, it's something to start with in
trying to re-implement the debugging demo under stackless.)