Make signal mask per-task
The "signal mask" in R3-Alpha paralleled the unix signal mask, and
would track whether it was a good time to listen to HALT signals via
Ctrl-C, or allow Recycle, etc.  It wasn't particularly fleshed out.

Ren-C utilized the signal mask in an attempt to mark moments when the
Ctrl-C signal should not be heeded, e.g. during stackful API calls.
These API calls were used in order to process the response to the
Ctrl-C signal itself, e.g. testing for values returned that meant
"I was canceled".