Remove halt hook from console, move to main
Because the console implemented a lot of work to get executions to
be controlled in terms of their cancelability--as well as handle
interpretation of exit codes--it was reused by the boot process to
run any usermode code (like scripts specified on the command line).

This did not enable scenarios where a build might be made that had
command line processing but no console.

Hence tying the cancellation into the console extension is clearly
wrong.  This switches to an approach that assumes cancelability is
managed on a granularity of rebHalt(), where the cancels always happen
and it's just a matter of whether a signal mask is there to accept it.

There are more nuances involved, e.g. as to whether file read or
write operations are cancelable.  It's not clear exactly how the
Revolt-cancelability state would be mirrored in the signal masks for
operations like read() and write() to decide whether they are allowed
to be canceled or not.  But for now we will just assume that all
such operations are occuring in cancelable contexts.