Give ACTION!s their own REBFRM*
In the initial conception of FRAME!, the idea was that since only
one function could be running in a frame at a time--and that non
function REBFRM* wouldn't have varlists to be reified into an
ANY-CONTEXT!, that a single REBFRM* would be reused for all function
calls at that level.  If the function were reified through its varlist,
then that reification would tear off that varlist from the frame
and it would get a new one.

While this is a nice optimization, it has the negative effect that
there is no reified way to refer to a frame which is just stepping
through a block or group.  If those things cannot be referred to by
an invariant handle in usermode code, then it restricts the language
that can be used in the debugger to inquire about those frames and
ask for automation on their behalf.

In addition, switching the executor of a frame midstream is a concept
that may not be ideal.  If a frame could use the same executor then
it could provide more assurances.  Seeking efficiency might be done
some other way.

This is a commit that gives each action execution its own frame, which
had to make various tweaks in order to get the tests to pass.