Allow FRAME! to be constructed for non-actions
This adds capability of frames that are not action frames to be made,
just to mark a place in an evaluation sequence.

It's a very primitive work-in-progress--but believed to be needed in
order to give usermode enough of a vocabulary to write an effective
debugger.  Also, it may hold the keys to making the "Executor" more
consistent...as it might be that a frame has a single executor over
its full execution.