Move CONSOLE to fully usermode implementation
The original goal of CONSOLE was to be able to remove itself from
the stack while usermode code ran (skin code or evaluation requests).
It was thus written in a somewhat convoluted way with callbacks.

Different philosophies about CONSOLE now have it making requests on
another stack via GO routines.  Given that design, the entire thing
can be moved to usermode.