Fix interoperability of CHAIN and YIELDER
Due to the way that CHAIN works, it was stacking pending calls onto
the frame started by a yielder that were not processed on the first
YIELD.  That broke one of the generator test variants, which was put
off for a later time.

That time is now.  This moves the stack in effect at the time the
yielder is called and puts that stack back onto the YIELD, which is
effectively the moment of the generator return.  It has an additional
clause on natural completion of the yielder's body.

There may be a more elegant way to do this--depending on how that
more elegant way would work with "goroutines".  Notes are added in
the soure about this, which may be pursued at a later date.