Make error backtraces limit to task boundaries
The current working of tasks is to leave the "main task" plugged into
the stack, and only unplug and replug tasks launched with GO.  This
meant that error traces showed more information than was desired,
since the errors climbed up to the top of the entire stack.

This limits the climb to roots of tasks when presenting error stacks.