Apply DEFER to the DO* helper routine
DO* is a usermode service function called to implement DO of types like
TEXT! or TAG!...things that are less foundational, and have some
fiddly preferences.  One issue is that when doing a script in a
directory, the directory was temporarily changed to that location--
and it needed to be changed back.

An attempt to catch QUITs was included, but it was somewhat messy
and did not account for FAIL.  DEFER provides a nicer option.