New_Expression_Executor() into Evaluator_Executor()
When stackless was started, it was difficult to entangle the very
monolithic evaluator to fit the new model.  Things like Argument
Fulfillment or Lookahead were broken out into their own independent
executors...because it wasn't clear how the Trampoline would handle
these components.

With time, the use of a STATE_BYTE() came to be standard practice in
natives.  Since it worked out well--and could be folded in efficiently
to the zeroing out of frame bits--it moved on to dispatchers and
executors also.

The idea of switching executors in the middle of a frame seemed less
appealing than having continuity.  Action_Executor() came to stick out
as a genuinely distinct frame category...and it was disentangled from
needing to interact with the Lookahead_Executor directly.

This takes a step toward reducing the total executor count by folding
the New_Expression_Executor() with the Reevaluator_Executor() to start
a generic Evaluator_Executor().