Move Parse_Group_Executor() into Parser_Executor()
The initial thought was that the granularity of execution wouldn't be
done with switch() statements of a single function, but with functions
that became identity entry points in a state machine for the stackless
trampoline--each yield to the trampoline constituting a hookable
debug point.

This concept moved to a more conventional notion of switching on a
byte of state, that would live in the "hot" header bits of a frame.
pushing new frames was almost always done if a continuation was
to be invoked.  If "hooks" were to be exposed (e.g. for debugging at
a step beyond just "each potentially recursive evaluation", the
granularity would have to come from another mechanism.

Pursuant to that, this merges the group executor step of parse in
with a parse STATE_BYTE mode...which was already partially the case
anyway.