Eliminate Finished_Executor(), make f->out return final
This pushes further the idea of re-integrating the evaluator.  It also
makes it so that returning a frame's output cell instead of
R_CONTINUATION is an implicit statement of completion, so that the
executor does not need to be responsible for setting it to nullptr or
anything else (it is trashed).