Identify Is_Action_Frame() by executor
Historically action frames were identified by when f->original was
not nullptr.  This uses the Action_Executor() as the way of telling.

Because executors are no longer nulled out when the final return of
a frame is done, this means that the GC step needs to be moved to
before the executor is called.  Otherwise, the un-dropped frame of
an action would be GC processed as if it were still alive...since
the action executor would be set.