Move Action_Executor() specific frame fields to union
This change moves toward an acknowledgment that different frames have
different needs for data.  Currently there are:

* Evaluator frames (blocks/va_lists)
* Action frames
* Parse frames
* Compose frames
* Scanner frames

What these all have in common is that they recurse, and that rather
than recurse directly via C they `return` and ask for that recursion
to be handled by a generic "Trampoline" function.

At present, Evaluator/Action/Parse/Compose also share in common an
"instruction" feed of Rebol array data, which is held locked and
advanced upon.  This may be useful for other recursive concepts like
Copy or Mold.  The scanner may come to use this as well.

However, the actual state that frames hold may vary.  Actions are
cooncerned with advancing parameters and filling in arguments with
specialized data...and this is only relevant to Action_Executor().

This moves the action-specific fields into the union inside the frame.
It's a considerable change, and in order to mitigate the effects a
convention for addressing the union fields is adopted which makes
macros for `f_xxx` look for a local variable called `f` and pick the
fields out, as conveniently as if it were still `f->xxx`.  Since frames
are called `f` more often than not, this works all right.