Make top-level PATH!, GET-PATH!, SET-PATH! stackless
This is a first step at rethinking PATH! execution that makes the
path evaluation at the top-level use R_CONTINUATION to evaluate groups.

(It breaks "subcell addressing" which was done in GOB! and STRUCT!,
which were bad hacks and need to be done another way.)

Path evaluations in the main Evaluator_Executor() are also changed to
be run through continuations--leveraging the new machinery.  Due to
the commonality between the PATH!, GET-PATH!, and SET-PATH! cases
they are run through a common `process-path`.