Push_Frame() passes executor, incremental cleanup
This tries to put a bit of order into the moments that the f->executor
is assigned or can be changed, by making it part of the Push_Frame()
in most cases.  Those places where it doesn't fit can use a
Push_Frame_Core() which leaves it as trash, and that INIT_F_EXECUTOR()
initializes it later.

The current concept is that after every frame completion, Trampoline
trashes the executor.  So stepwise re-use of frames has to put the
executor back each time.  This seems best, since TRAMPOLINE_KEEPALIVE
could otherwise leave a frame on the stack that would just keep
evaluating... but it may be that future optimizations choose a
different strategy.

As part of the cleanup, some of the touched code areas were using the
SHORTHAND utility.  This has been mostly supplanted by the `f_xxx`
macros, and is confusing enough to eliminate.  Also, it hasn't been
clear that compiler optimizations was removing the trick fully.