Update evaluator comment blocks
The evaluator had some comment blocks that were all the way to the
left for many lines...which is a place usually reserved for top-level
declarations.  It seemed good at some point in time, but now it looks
to be a bit heavy-handed.

This indents those blocks, and while doing so gives the comment blocks
a refresh of editing so they are more current.