Make some REBTYPE() dispatchers a bit clearer
There was a trend to make generic dispatchers use `v` to mean "value
being dispatched", which wasn't particularly clear...and becomes less
clear as dispatch with ARG(value) is coded for.

This is some residue from a failed experiment that adjusts some of
those names in a way that is likely clearer.  Also breaks out the
use of `sym` as a standard for the symbol of the `verb` of the generic.