Make CASE evaluate into spare cell
In the first attempts to convert natives to act stacklessly, the
question came up of whether frames would always be forced to evaluate
into the output cell of their parents, or if other locations could
be chosen.

This changes CASE back to evaluate into the spare cell, based on
the belief that this pattern is going to be solved.