Fix GC bugs in Make_Invocation_Frame() calls
The choice to identify action frames by having Action_Executor() meant
that once a frame is pushed, it would have to be valid enough to be
seen by the GC if any evaluations ran.  Cases for Make_Invocation_Frame
were pushing a frame just to capture the DSP position, prior to doing
an evaluation.  This was a bit heavy-handed when it could have just
captured the DSP--and led to some convoluted code for having to drop
the partially-constructed frame when it probably shouldn't exist at all

This tidies up those callsites to make the DSP capture a separate thing
and only create the frame once it is known to be needed...and can
be fully constructed prior to running another evaluation.