Move `v` and `gotten` into the frame
Frames initially had to be able to hold both an evaluator state as
well as an action state.  Now that actions get their own frames, there
is no need to have local variables for the `v` and `gotten`...they
can be in the evaluator specific union fields.

To help add clarity to what these are for, they are renamed.  It is
also the case that when continuity is possible across continuations,
these members do not need to be reassigned.

While it was feared that putting these into the frame might lower
performance, it turns out that there's negligible impact.