Initial implementation of "rich compare" executor
This begins the idea of trying to move toward a stackless concept of
compare, while modeling a comparison system after Python3's
"rich comparisons"

https://www.python.org/dev/peps/pep-0207/

It's only a sketch of how this might work.  It means that things like
EQUAL? are no longer natives, but specializations of a generic COMPARE
operator.

With no native EQUAL? the performance of the comparison-delegating
FOR loop would be worse than it was already by using string binding.
This sketches a way to deal with it by trying to make a comparison
operator delegation that may-or-may-not need to create a frame, so
that optimizations could avoid the overhead of continuations.  This
makes FOR somewhat more circuitous...requiring a frame state for
every comparison.  It will require review, but demonstrates a
functioning concept to look at.