Reduce_To_Stack_Throws() out of Core
This restyles a feature in PARSE which let you reduce a BLOCK! for
KEEP input to not use Reduce_To_Stack_Throws().  It uses CONTINUE_WITH
that calls the REDUCE native with the block to reduce.  (This may not
ultimately be desirable.)

Consequently, the only remaining piece of code that is using the
Reduce_To_Stack_Throws() is code in the FFI.  The FFI was originally
made by Atronix, and was written in C before good concepts of dividing
labor in extenssions between usermode and native code were established.
This moves the code there pending a rewrite of MAKE STRUCT! which would
better break things down to use far less C than it does today.