First cut implementation of escaped spellings
This is a rough concept of how "rare" casing variations could be
escaped, fitting in with the same mechanism used to make high levels
of quoting "pop out" the dta into an escaped cell.  The idea is that
spelling variations would use larger HEART_BYTE() values, with a
sufficiently high level implying that the cell points to an unbound
word with the full spelling.

Something that makes this approach difficult are that values which split
their information across two cells in this way have to sync their
kind byte and heart byte if they are changed.  Being able to share the
"unescaped" cell with other instances makes this problematic, and
require making a new copy.

Another questionable issue regarding sharing arises with virtual
binding.  There's no room for virtual binding information in the
escaped cell (of a high-numbered quoted or weird spelling).  This
means the information goes into the unescaped cell--which can be shared.
If a copy is not made, then words with strange spellings (or
for that matter, high quote levels) may be shared across contexts
where their virtual binding data needs to be different, resulting in
competition for the cache data.  This is a performance problem--and
presumably a rare one--but worth keeping in mind.